Change Data Type
Change Data Type

+++ GenericOpenAPI/Models/RealisasiPendapatan.cs
@@ -4,6 +4,6 @@
     {
         public string KdPer { get; set; }
         public string NmPer { get; set; }
-        public float Nilai { get; set; }
+        public decimal Nilai { get; set; }
     }
 }
